feat(ssg): :art: decoupling postprocessor
